refactor: core 패키지 삭제 및 클래스 이동
