+++ tomcat/src/main/java/org/apache/catalina/servlets/DefaultServlet.java
@@ -7,8 +7,8 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.apache.catalina.core.request.HttpRequest;
-import org.apache.catalina.core.response.HttpResponse;
+import org.apache.catalina.servlets.http.request.HttpRequest;
+import org.apache.catalina.servlets.http.response.HttpResponse;
 
 /**
  * 정적 리소스를 관리하는 서블릿입니다.
