+++ tomcat/src/main/java/org/apache/catalina/servlets/Servlet.java
@@ -1,8 +1,8 @@
 package org.apache.catalina.servlets;
 
 import java.io.IOException;
-import org.apache.catalina.core.request.HttpRequest;
-import org.apache.catalina.core.response.HttpResponse;
+import org.apache.catalina.servlets.http.request.HttpRequest;
+import org.apache.catalina.servlets.http.response.HttpResponse;
 
 public interface Servlet {
     void service(HttpRequest request, HttpResponse response) throws IOException;
