+++ tomcat/src/main/java/org/apache/catalina/servlets/http/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.core.response;
+package org.apache.catalina.servlets.http;
 
 public class Cookie {
