+++ tomcat/src/main/java/org/apache/catalina/servlets/http/response/HttpResponse.java
@@ -1,7 +1,8 @@
-package org.apache.catalina.core.response;
+package org.apache.catalina.servlets.http.response;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
+import org.apache.catalina.servlets.http.Cookie;
 
 public class HttpResponse {
