+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -3,8 +3,8 @@ package org.apache.coyote.http11;
 import com.techcourse.exception.UncheckedServletException;
 import java.io.IOException;
 import java.net.Socket;
-import org.apache.catalina.core.request.HttpRequest;
-import org.apache.catalina.core.response.HttpResponse;
+import org.apache.catalina.servlets.http.request.HttpRequest;
+import org.apache.catalina.servlets.http.response.HttpResponse;
 import org.apache.catalina.mapper.Mapper;
 import org.apache.catalina.servlets.Servlet;
 import org.apache.coyote.Processor;
